+++ packages/server/src/tools/glob.ts
@@ -29,7 +29,7 @@ export interface GlobToolParams {
 /**
  * Implementation of the Glob tool logic (moved from CLI)
  */
-export class GlobLogic extends BaseTool<GlobToolParams, ToolResult> {
+export class GlobTool extends BaseTool<GlobToolParams, ToolResult> {
   static readonly Name = 'glob'; // Keep static name
 
   /**
@@ -43,9 +43,9 @@ export class GlobLogic extends BaseTool<GlobToolParams, ToolResult> {
    */
   constructor(rootDirectory: string) {
     super(
-      GlobLogic.Name,
-      '', // Display name handled by CLI wrapper
-      '', // Description handled by CLI wrapper
+      GlobTool.Name,
+      'FindFiles', // Display name handled by CLI wrapper
+      'Efficiently finds files matching specific glob patterns (e.g., `src/**/*.ts`, `**/*.md`), returning absolute paths sorted by modification time (newest first). Ideal for quickly locating files based on their name or path structure, especially in large codebases.', // Description handled by CLI wrapper
       {
         properties: {
           pattern: {
