+++ packages/server/src/utils/BackgroundTerminalAnalyzer.ts
